tests/wpt/web-platform-tests/import-maps/module-map-key.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="importmap">
{
  "imports": {
    "./resources/log.js?pipe=sub&name=A": "./resources/log.js?pipe=sub&name=B"
  }
}
</script>
<script>
const log = [];

promise_test(() => {
  return import("./resources/log.js?pipe=sub&name=A")
    .then(() => import("./resources/log.js?pipe=sub&name=B"))
    .then(() => assert_array_equals(log, ["log:B"]))
  },
  "Module map's key is the URL after import map resolution");
</script>
